+++ index.html
@@ -168,7 +168,7 @@
     <div id="hidden_menu">
         <div class="menu">
             <div class="header">
-                menu<i class="fa fa-times-circle-o" aria-hidden="true"></i>
+                menu<i class="fa fa-times-circle-o close-btn" aria-hidden="true"></i>
             </div>
             <div class="menu-container">
                 <div class="menu-body">
